add Id type

+++ src/main/java/org/atlasapi/ChannelAvailableOnSettingTask.java
@@ -46,7 +46,7 @@ public class ChannelAvailableOnSettingTask extends ScheduledTask {
             
             addAll(availableOn, filter(transform(channel.getAliases(), TO_AVAILABLE_ON), notNull()));
             
-            collection.update(where().idEquals(channel.getId()).build(), update().setField(ChannelTranslator.AVAILABLE_ON, availableOn).build());
+            collection.update(where().idEquals(channel.getId().longValue()).build(), update().setField(ChannelTranslator.AVAILABLE_ON, availableOn).build());
         }
     }
